Add workshop confirm delete page and functionality

+++ workshops/views.py
@@ -77,18 +77,39 @@ def edit_workshop(request, workshop_id):
     return render(request, template, context)
 
 
+# @login_required
+# def delete_workshop(request, workshop_id):
+#     """ Delete a workshop in the store """
+#     if not request.user.is_superuser:
+#         messages.error(request, 'Sorry, only shop owners can do that.')
+#         return redirect(reverse('home'))
+        
+#     workshop = get_object_or_404(Workshops, pk=workshop_id)
+#     workshop.delete()
+#     messages.success(request, 'Workshop deleted')
+
+#     return redirect(reverse('workshops'))
+
+
 @login_required
 def delete_workshop(request, workshop_id):
-    """ Delete a workshop in the store """
+    """ Delete a workshop """
     if not request.user.is_superuser:
-        messages.error(request, 'Sorry, only shop owners can do that.')
+        messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
-        
-    workshop = get_object_or_404(Workshops, pk=workshop_id)
-    workshop.delete()
-    messages.success(request, 'Workshop deleted')
 
-    return redirect(reverse('workshops'))
+    workshop = get_object_or_404(Workshops, pk=workshop_id)
+    if request.method == 'POST':
+        workshop.delete()
+        messages.success(request, 'Workshop deleted successfully.')
+        return redirect(reverse('workshops'))
+
+    template = 'workshops/workshop_confirm_delete.html'
+    context = {
+        'workshop': workshop,
+    }
+    
+    return render(request, template, context)
 
 
 def WorkshopRequest(request):
